refactor(fields): remove unused `MultiSelectField` component and update related layout logic
Removed the obsolete `MultiSelectField` implementation. Updated `ConfigPanel` to handle cached restrictions for improved validation consistency during delay gaps. Extended tests for better coverage of enum list types and ensured compatibility with revised `value-type-parser` enhancements.
