+++ frontend/src/components/base/fields/fields/EnumListField.tsx
@@ -8,38 +8,45 @@
  * does it submit to any jurisdiction.
  */
 
-import { Check, ChevronDown, X } from 'lucide-react'
 import { useTranslation } from 'react-i18next'
-import { GlyphFieldWrapper } from './GlyphFieldWrapper'
-import { Badge } from '@/components/ui/badge'
-import { Button } from '@/components/ui/button'
 import {
-  Popover,
-  PopoverContent,
-  PopoverTrigger,
-} from '@/components/ui/popover'
-import { cn } from '@/lib/utils'
+  Combobox,
+  ComboboxChip,
+  ComboboxChips,
+  ComboboxChipsInput,
+  ComboboxContent,
+  ComboboxEmpty,
+  ComboboxItem,
+  ComboboxList,
+  ComboboxValue,
+  useComboboxAnchor,
+} from '@/components/ui/combobox'
+import { useFieldErrors } from '@/features/fable-builder/context/BlockValidationContext'
 
 export interface EnumListFieldProps {
   id: string
   configKey: string
+  /** Comma-separated, matching the `list[str]` wire encoding. */
   value: string
   onChange: (value: string) => void
-  options: Array<string>
+  options: ReadonlyArray<string>
+  /** `list[enumClosed[…]]` (closed, default) vs `list[enum[…]]` (open;
+   * free-form not wired yet, kept for forward compat). */
+  closed?: boolean
   placeholder?: string
   disabled?: boolean
   className?: string
 }
 
-function parseListValue(value: string): Array<string> {
-  if (!value.trim()) return []
-  return value
+function parseListValue(raw: string): Array<string> {
+  if (!raw.trim()) return []
+  return raw
     .split(',')
-    .map((item) => item.trim())
+    .map((s) => s.trim())
     .filter(Boolean)
 }
 
-function serializeListValue(items: Array<string>): string {
+function serializeListValue(items: ReadonlyArray<string>): string {
   return items.join(',')
 }
 
@@ -51,130 +58,70 @@ export function EnumListField({
   options,
   placeholder,
   disabled,
-  className,
 }: EnumListFieldProps) {
   const { t } = useTranslation('common')
-  const resolvedPlaceholder = placeholder ?? t('field.selectPlaceholder')
+  const resolvedPlaceholder = placeholder ?? t('field.addItemPlaceholder')
+  const items = parseListValue(value)
+  const anchor = useComboboxAnchor()
+  // Hide persisted values no longer in `options`; storage keeps them until
+  // the next edit (we don't fire onChange on render).
+  const validItems = items.filter((v) => options.includes(v))
+
+  // No GlyphFieldWrapper: glyph mode is meaningless for multi-select, and
+  // its InputGroup chrome fights the Combobox's own chip container. Render
+  // errors inline instead.
+  const fieldErrors = useFieldErrors()?.[configKey] ?? null
+  const hasFieldError = fieldErrors !== null && fieldErrors.length > 0
+  const errorMessage = hasFieldError
+    ? fieldErrors.length > 1
+      ? `${fieldErrors[0]} (+${fieldErrors.length - 1} more)`
+      : fieldErrors[0]
+    : null
 
   return (
-    <GlyphFieldWrapper
-      id={id}
-      configKey={configKey}
-      value={value}
-      onChange={onChange}
-      placeholder={resolvedPlaceholder}
-      disabled={disabled}
-      className={className}
-      allowGlyphMode={false}
-    >
-      <EnumListFieldConcrete
-        id={id}
-        value={value}
-        onChange={onChange}
-        options={options}
-        placeholder={resolvedPlaceholder}
+    <div>
+      <Combobox<string, true>
+        multiple
+        autoHighlight
+        items={[...options]}
+        value={validItems}
+        onValueChange={(next) => onChange(serializeListValue(next))}
         disabled={disabled}
-        className={className}
-      />
-    </GlyphFieldWrapper>
-  )
-}
-
-function EnumListFieldConcrete({
-  id,
-  value,
-  onChange,
-  options,
-  placeholder,
-  disabled,
-  className,
-}: Omit<EnumListFieldProps, 'configKey'>) {
-  const { t } = useTranslation('common')
-  const selected = parseListValue(value)
-  const selectedSet = new Set(selected)
-  const resolvedPlaceholder = placeholder ?? t('field.selectPlaceholder')
-  const buttonText =
-    selected.length > 0 ? selected.join(', ') : resolvedPlaceholder
-
-  function toggleOption(option: string): void {
-    const nextSet = new Set(selectedSet)
-    if (nextSet.has(option)) {
-      nextSet.delete(option)
-    } else {
-      nextSet.add(option)
-    }
-    onChange(serializeListValue(options.filter((item) => nextSet.has(item))))
-  }
-
-  function removeOption(option: string): void {
-    onChange(serializeListValue(selected.filter((item) => item !== option)))
-  }
-
-  return (
-    <div className="min-w-0 space-y-2">
-      <Popover>
-        <PopoverTrigger
-          render={
-            <Button
-              id={id}
-              type="button"
-              variant="outline"
-              className={cn('h-9 w-full justify-between px-3', className)}
-              disabled={disabled}
-            />
-          }
+      >
+        <ComboboxChips
+          ref={anchor}
+          className={hasFieldError ? 'border-destructive' : undefined}
         >
-          <span
-            className={cn(
-              'truncate',
-              selected.length === 0 && 'text-muted-foreground',
+          <ComboboxValue>
+            {(values: Array<string>) => (
+              <>
+                {values.map((v) => (
+                  <ComboboxChip key={v}>{v}</ComboboxChip>
+                ))}
+                <ComboboxChipsInput
+                  id={id}
+                  placeholder={
+                    values.length === 0 ? resolvedPlaceholder : undefined
+                  }
+                  disabled={disabled}
+                />
+              </>
             )}
-          >
-            {buttonText}
-          </span>
-          <ChevronDown className="h-4 w-4 opacity-50" />
-        </PopoverTrigger>
-        <PopoverContent className="max-h-72 w-(--anchor-width) gap-1 overflow-y-auto p-1">
-          {options.map((option) => {
-            const checked = selectedSet.has(option)
-            return (
-              <button
-                key={option}
-                type="button"
-                className={cn(
-                  'flex w-full items-center gap-2 rounded-sm px-2 py-1.5 text-left text-sm',
-                  'outline-none hover:bg-accent hover:text-accent-foreground',
-                )}
-                onClick={() => toggleOption(option)}
-              >
-                <span className="flex h-4 w-4 items-center justify-center">
-                  {checked && <Check className="h-4 w-4" />}
-                </span>
-                <span className="min-w-0 flex-1 truncate">{option}</span>
-              </button>
-            )
-          })}
-        </PopoverContent>
-      </Popover>
-
-      {selected.length > 0 && (
-        <div className="flex min-w-0 flex-wrap gap-1.5">
-          {selected.map((item) => (
-            <Badge key={item} variant="secondary" className="gap-1 pr-1">
-              {item}
-              {!disabled && (
-                <button
-                  type="button"
-                  onClick={() => removeOption(item)}
-                  className="ml-1 rounded-full p-0.5 transition-colors hover:bg-muted-foreground/20"
-                  aria-label={t('removeTag', { tag: item })}
-                >
-                  <X className="h-3 w-3" />
-                </button>
-              )}
-            </Badge>
-          ))}
-        </div>
+          </ComboboxValue>
+        </ComboboxChips>
+        <ComboboxContent anchor={anchor}>
+          <ComboboxEmpty>{t('field.noMatches')}</ComboboxEmpty>
+          <ComboboxList>
+            {(item: string) => (
+              <ComboboxItem key={item} value={item}>
+                {item}
+              </ComboboxItem>
+            )}
+          </ComboboxList>
+        </ComboboxContent>
+      </Combobox>
+      {errorMessage && (
+        <p className="mt-1 truncate text-xs text-destructive">{errorMessage}</p>
       )}
     </div>
   )
